04 Identity Server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -44,13 +45,15 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction([FromBody] CreateAuctionDto auctionDto)
     {
         var auction = mapper.Map<Auction>(auctionDto);
 
         // TODO: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity?.Name ?? throw new InvalidOperationException("User not found");
+
         context.Auctions.Add(auction);
 
         //CREATE CONSUMER
@@ -68,6 +71,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult<AuctionDto>> UpdateAuction(Guid id, [FromBody] UpdateAuctionDto auctionDto)
     {
@@ -81,6 +85,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         }
 
         // TODO: check seller is the same as current user
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         //updating props
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
@@ -101,6 +106,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         return Ok(mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -111,10 +117,11 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
         }
 
         // TODO: check seller is the same as current user
+        if (auction.Seller != User.Identity?.Name) return Forbid();
         context.Auctions.Remove(auction);
 
         await publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
-        
+
         var result = await context.SaveChangesAsync() > 0;
         if (!result)
         {
